Recurring quests only purge completed quests on rollover

+++ GameServer/ECS-Services/DailyQuestService.cs
@@ -83,7 +83,8 @@ public class DailyQuestService
 
             foreach (var existingDailyQuest in existingDailyQuests)
             {
-                GameServer.Database.DeleteObject(existingDailyQuest);
+                if(existingDailyQuest.Step <= -1)
+                    GameServer.Database.DeleteObject(existingDailyQuest);
             }
